+++ src/Adapter/Application/Application.mjs
@@ -1,6 +1,5 @@
-/** @typedef {import("../../../../flux-studis-selfservice-frontend/src/Adapter/Page/PAGE.mjs").PAGE} PAGE */
 /** @typedef {import("../../../../flux-studis-selfservice-frontend/src/Adapter/Post/Post.mjs").Post} Post */
 
 /**
- * @typedef {{"identification-number": string, page: PAGE, posts: Post[]}} Application
+ * @typedef {{"identification-number": string, page: string, posts: Post[]}} Application
  */