+++ src/Service/Data/Command/GetIntendedDegreeProgram2Command.mjs
@@ -1,0 +1,50 @@
+import INTENDED_DEGREE_PROGRAM_2 from "../../../Adapter/Data/IntendedDegreeProgram2/intended-degree-program-2.json" assert {type: "json"};
+
+/** @typedef {import("../../../../../flux-studis-selfservice-frontend/src/Adapter/IntendedDegreeProgram/ChosenIntendedDegreeProgram.mjs").ChosenIntendedDegreeProgram} ChosenIntendedDegreeProgram */
+/** @typedef {import("../../../../../flux-studis-selfservice-frontend/src/Adapter/IntendedDegreeProgram2/ChosenIntendedDegreeProgram2.mjs").ChosenIntendedDegreeProgram2} ChosenIntendedDegreeProgram2 */
+/** @typedef {import("../Port/DataService.mjs").DataService} DataService */
+/** @typedef {import("../../../../../flux-studis-selfservice-frontend/src/Adapter/IntendedDegreeProgram2/IntendedDegreeProgram2.mjs").IntendedDegreeProgram2} IntendedDegreeProgram2 */
+
+export class GetIntendedDegreeProgram2Command {
+    /**
+     * @type {DataService}
+     */
+    #data_service;
+
+    /**
+     * @param {DataService} data_service
+     * @returns {GetIntendedDegreeProgram2Command}
+     */
+    static new(data_service) {
+        return new this(
+            data_service
+        );
+    }
+
+    /**
+     * @param {DataService} data_service
+     * @private
+     */
+    constructor(data_service) {
+        this.#data_service = data_service;
+    }
+
+    /**
+     * @param {ChosenIntendedDegreeProgram} chosen_intended_degree_program
+     * @param {ChosenIntendedDegreeProgram2 | null} values
+     * @returns {Promise<IntendedDegreeProgram2>}
+     */
+    async getIntendedDegreeProgram2(chosen_intended_degree_program, values = null) {
+        const subject = (await this.#data_service.getSubjects()).find(_subject => _subject.id === chosen_intended_degree_program.subject);
+
+        const _subject = structuredClone(subject);
+        delete _subject.combinations;
+
+        return {
+            ...INTENDED_DEGREE_PROGRAM_2,
+            subject: _subject,
+            combination: subject.combinations.find(combination => combination.id === chosen_intended_degree_program.combination),
+            values
+        };
+    }
+}