+++ CHANGELOG.md
@@ -6,11 +6,17 @@ Changes:
 
 \-
 
+## v2023-01-20-1
+
+Changes:
+
+- Split to commands
+
 ## v2023-01-18-1
 
 Changes:
 
-\-
+- `HttpRequest` / `HttpResponse`
 
 ## v2023-01-17-2